% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{As2Vs}
\alias{As2Vs}
\title{An Rcpp function that calculates the bootstapped Vs}
\usage{
As2Vs(AsByB, V)
}
\arguments{
\item{V}{left singular vectors of original sample data}

\item{AsByK}{number of bootstrap samples indexed by K}
}
\value{
A list of bootstrapped Vs indexed by Bootstrap samples
}
\description{
An Rcpp function that calculates the bootstapped Vs
}
