\name{biostat815project-package}
\alias{biostat815project-package}
\alias{biostat815project}
\docType{package}
\title{
\packageTitle{biostat815project}
}
\description{
\packageDescription{biostat815project}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{biostat815project}
\packageIndices{biostat815project}
}
\author{
Chao Gao
\packageAuthor{biostat815project}

Maintainer: \packageMaintainer{biostat815project}
}
\references{
Fast, Exact Bootstrap Principal Component Analysis for p > 1 million. (Fisher et al, 2016)
}
\keyword{ package }
\seealso{
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
results <- bootPCA(Y,3,50,0.05)
}
