% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bootSVD_LD}
\alias{bootSVD_LD}
\title{An Rcpp function that calculates the bootstrap distribution of the principal components (PCs) of a low dimensional matrix.}
\usage{
bootSVD_LD(n, DUt, K, B, centerSamples = 0L)
}
\arguments{
\item{n}{A (n x n) matrix of sample scores (i.e. the product of singular values and transpose(right singular vectors)}

\item{DUt}{transpose of UD (matrix of scores, were rows denote individuals, and columns denote measurements in the PC space.)}

\item{K}{the number of PCs to be estimated.}

\item{B}{the number of bootstrap samples}

\item{centerSamples}{whether each bootstrap sample should be centered before calculating the SVD, 1=Yes, 0=NO(default).}
}
\value{
A list containing left, right singular vectors and singular values of SVD(sample scores)
}
\description{
An Rcpp function that calculates the bootstrap distribution of the principal components (PCs) of a low dimensional matrix.
}
