% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{genBootIndices}
\alias{genBootIndices}
\title{An Rcpp function that generate bootstrapped indices}
\usage{
genBootIndices(B, n)
}
\arguments{
\item{B}{number of bootstrap samples}

\item{n}{number of original sample}
}
\value{
A (B x n) matrix  containing the bootstrapped indices
}
\description{
An Rcpp function that generate bootstrapped indices
}
