% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getHDpercentiles}
\alias{getHDpercentiles}
\title{An Rcpp function that calculate bootstrap percentile-based confidence intervals for the PCs of original data.}
\usage{
getHDpercentiles(VsByK, B, K, p, alpha)
}
\arguments{
\item{VsByK}{A list of Vs indexed by K}

\item{B}{number of bootstrap samples}

\item{K}{first K PCs of interest}

\item{p}{number of features}

\item{alpha}{(100*alpha/2, 100(1-alpha/2)) CI}
}
\value{
A list of bootstrap percentile-based confidence intervals for the PCs.
}
\description{
An Rcpp function that calculate bootstrap percentile-based confidence intervals for the PCs of original data.
}
