% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getLDpercentiles}
\alias{getLDpercentiles}
\title{An Rcpp function that calculate bootstrap percentile-based confidence intervals for the PCs of sample socores.}
\usage{
getLDpercentiles(AsByK, B, K, n, alpha)
}
\arguments{
\item{B}{number of bootstrap samples}

\item{K}{first K PCs of interest}

\item{alpha}{(100*alpha/2, 100(1-alpha/2)) CI}

\item{VsByK}{A list of Vs indexed by K}

\item{p}{number of features}
}
\value{
A list of bootstrap percentile-based confidence intervals for the PCs.
}
\description{
An Rcpp function that calculate bootstrap percentile-based confidence intervals for the PCs of sample socores.
}
