% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMomentsAndMomentCI}
\alias{getMomentsAndMomentCI}
\title{An Rcpp function that calculate bootstrap moments and moment-based confidence intervals for the PCs.}
\usage{
getMomentsAndMomentCI(AsByK, V, B, K, n)
}
\arguments{
\item{AsByK}{number of bootstrap samples indexed by K}

\item{V}{left singular vectors of original sample data}

\item{B}{number of bootstrap samples}

\item{K}{first K PCs of interest}

\item{n}{number of subjects}
}
\value{
A list of moments and moment-based CIs for the PCs
}
\description{
An Rcpp function that calculate bootstrap moments and moment-based confidence intervals for the PCs.
}
