% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reindexMatricesByK}
\alias{reindexMatricesByK}
\title{An Rcpp function that Used for calculation of low dimensional standard errors & percentiles, by re-indexing the \eqn{A^b} by PC index (\eqn{k}) rather than bootstrap index (\eqn{b}).}
\usage{
reindexMatricesByK(AsByB, B, K)
}
\arguments{
\item{AsByB}{a size B list of matrices in the order boostrap samples}

\item{B}{number of boostrap samples}

\item{K}{first K PCs of interest}
}
\value{
A list of reordered matrices (order by PCs)
}
\description{
An Rcpp function that Used for calculation of low dimensional standard errors & percentiles, by re-indexing the \eqn{A^b} by PC index (\eqn{k}) rather than bootstrap index (\eqn{b}).
}
